pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "schedule>=1.2.2",
    "slack-sdk>=3.34.0",
]
